refactor: moved version handlers to version-repo project
